refactoring the code..
Drop the Git Tag if the build has failed.

Change is in ReleaseEnvironment#finalizeSCMRepo step (1).
